organise code
