+++ src/config.ts
@@ -1,1 +1,2 @@
 export const appName: string = 'SchoolMoney';
+export const baseApiUrl: string = 'http://localhost:8000';
