+++ src/lib/api/auth.ts
@@ -1,27 +1,8 @@
-import { Token } from '$lib/models/auth';
+import { type Token, type Session } from '$lib/models/auth';
 import { jwtDecode } from 'jwt-decode';
+import { baseApiUrl } from '../../config';
 
 const TOKEN_KEY = 'access-token';
-const API_URL = 'http://localhost:8000';
-
-export enum Privilege {
-  STANDARD_USER = 1,
-  ADMIN_USER = 10,
-}
-
-export enum Status {
-  DISABLED = 0,
-  LOCKED = 1,
-  ENABLED = 10,
-}
-
-export type Session = {
-  user_id: number;
-  username: string;
-  email?: string;
-  privilege: Privilege;
-  status: Status;
-};
 
 export function getToken() {
 	return localStorage.getItem(TOKEN_KEY);
@@ -45,7 +26,7 @@ function setToken(token: string) {
 }
 
 export async function login(username: string, password: string) {
-	const resp = await fetch(`${API_URL}/api/auth/login`, {
+	const resp = await fetch(`${baseApiUrl}/api/auth/login`, {
 		method: 'POST',
 		headers: {
 			'Content-Type': 'application/json'
@@ -56,9 +37,8 @@ export async function login(username: string, password: string) {
 		const { detail } = await resp.json();
 		throw new Error(detail);
 	}
-	const { access_token, token_type, expires, refresh_token } = await resp.json();
-	let token = new Token(access_token, token_type, expires, refresh_token);
-	setToken(token.access_token);
+	const { access_token }: Token = await resp.json() as Token;
+	setToken(access_token);
 }
 
 export async function getUserDetails() {
@@ -66,7 +46,7 @@ export async function getUserDetails() {
 	if (!token) {
 		throw new Error('User is not authenticated');
 	}
-	const resp = await fetch(`${API_URL}/api/user/me`, {
+	const resp = await fetch(`${baseApiUrl}/api/user/me`, {
 		method: 'GET',
 		headers: {
 			Authorization: `Bearer ${token}`,
@@ -83,7 +63,7 @@ export async function logout() {
 	if (!token) {
 		throw new Error('User is not authenticated');
 	}
-	const resp = await fetch(`${API_URL}/api/auth/logout`, {
+	const resp = await fetch(`${baseApiUrl}/api/auth/logout`, {
 		method: 'POST',
 		headers: {
 			Authorization: `Bearer ${token}`,
@@ -99,7 +79,7 @@ export async function logout() {
 }
 
 export async function register(username: string, password: string) {
-	const resp = await fetch(`${API_URL}/api/user/register`, {
+	const resp = await fetch(`${baseApiUrl}/api/user/register`, {
 		method: 'POST',
 		headers: {
 			'Content-Type': 'application/json'
