+++ src/lib/models/auth.ts
@@ -1,13 +1,25 @@
-export class Token {
-    access_token: string;
-    token_type: string;
-    expires: number;
-    refresh_token: string;
-
-    constructor(access_token: string, token_type: string, expires: number, refresh_token: string) {
-        this.access_token = access_token;
-        this.token_type = token_type;
-        this.expires = expires;
-        this.refresh_token = refresh_token;
-    }
+export type Token = {
+  access_token: string;
+  token_type: string;
+  expires: number;
+  refresh_token: string;
 }
+
+export enum Privilege {
+  STANDARD_USER = 1,
+  ADMIN_USER = 10,
+}
+
+export enum Status {
+  DISABLED = 0,
+  LOCKED = 1,
+  ENABLED = 10,
+}
+
+export type Session = {
+  user_id: number;
+  username: string;
+  email?: string;
+  privilege: Privilege;
+  status: Status;
+};
