+++ src/routes/+layout.svelte
@@ -1,2 +1,1 @@
 <slot></slot>
-
